feat: Implemented MediatR

+++ API/Middleware/ExceptionMiddleware.cs
@@ -35,9 +35,9 @@ namespace API.Middleware
                 context.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
 
                 var response = _env.IsDevelopment()
-                    ? new ApiException((int) HttpStatusCode.InternalServerError, 
+                    ? new ApiException(HttpStatusCode.InternalServerError, 
                     ex.Message, ex.StackTrace.ToString())
-                    : new ApiException((int)HttpStatusCode.InternalServerError);
+                    : new ApiException(HttpStatusCode.InternalServerError);
 
                 var options = new JsonSerializerOptions{PropertyNamingPolicy =
                 JsonNamingPolicy.CamelCase};
